Refactor `rudr apps`
Move some basic functions to cmd.util.helpers.go and
decompose some of them for further calling, like
`rudr unbind`
